backend/requirements.txt
annotated-types==0.6.0
anyio==4.9.0
asgiref==3.8.1
beautifulsoup4==4.12.3
blinker==1.8.2
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
currencyapicom==0.1.1
cymem==2.0.8
distro==1.9.0
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
everapi==0.1.1
exceptiongroup==1.2.2
fastapi==0.115.12
Flask==3.0.3
gurobipy==11.0.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.3.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
insta-scrape==2.1.2
instagrapi==2.1.1
instascrape==0.1
itsdangerous==2.2.0
Jinja2==3.1.3
jiter==0.9.0
joblib==1.3.2
langcodes==3.3.0
lxml==5.1.0
MarkupSafe==2.1.5
murmurhash==1.0.10
nltk==3.8.1
numpy==1.26.4
openai==1.68.2
packaging==24.0
pandas==2.2.1
pillow==10.3.0
preshed==3.0.9
psycopg==3.1.18
psycopg-binary==3.1.18
pycryptodomex==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
PyQt5==5.15.9
pyqt5-plugins @ file:///C:/Users/chams/Downloads/pyqt5_plugins-5.15.4.2.2-py3-none-win_amd64.whl#sha256=1d9acf423062c6694a5e68782fb54ce90355481d33281c9193da45c43426853c
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyqt5-tools==5.15.9.3.3
PySocks==1.7.1
pyswip @ git+https://github.com/ML-KULeuven/pyswip@20a3fc841adfa7e5768e696d21b3e228dbe11e07
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
regex==2023.12.25
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.4.8
starlette==0.46.1
thinc==8.2.3
tqdm==4.66.2
typer==0.9.4
typing_extensions==4.10.0
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.2.1
uvicorn==0.34.0
validate_email==1.3
wasabi==1.1.2
weasel==0.3.4
Werkzeug==3.0.3
